+++ packages/types/src/index.ts
@@ -5,3 +5,4 @@ export * from "./api/users";
 export * from "./api/activity";
 export * from "./api/moderation";
 export * from "./api/feeds";
+export * from "./domain/submission";
